backend/requirements.txt
cassandra-driver==3.20.1
Flask==1.1.1
flask-socketio==4.2.1
eventlet==0.25.1
Flask-RESTful==0.3.7
Flask-Cors==3.0.8